image-server/requirements.txt
fastapi
uvicorn
deta
python-multipart
python-dotenv
